third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-1/case-sensitive-suggestions-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that text prompt suggestions' casing follows that of the user input.


Running: testForUpperCase

Running: testForLowerCase

Running: testForMixedCase
